Copied AbstractContractDetails into AionContractDetailsImpl:
 - removed inheritance from AbstractContractDetails;
 - deleted abstract class.
